scripts/域名到期监控/requirements.txt

flask==2.3.2
prometheus-client==0.17.0
pyyaml==6.0.1
whois==0.9.27
APScheduler==3.11.0